LayoutTests/fast/webgpu/ftoi-expected.txt

-65504,-2147483648,65504,2147483520,0,0,-65504,65504,0,0,-2147483648,2147483520,-65504,-2147483648,65504,2147483520,0,0,-65504,65504,0,0,-2147483648,2147483520
0,0,65504,4294967040,0,0,0,65504,0,0,0,4294967040,0,0,65504,4294967040,0,0,0,65504,0,0,0,4294967040
PASS output[0] is -65504
PASS output[1] is -2147483648
PASS output[2] is 65504
PASS output[3] is 2147483520
PASS output[4] is 0
PASS output[5] is 0
PASS output[6] is -65504
PASS output[7] is 65504
PASS output[8] is 0
PASS output[9] is 0
PASS output[10] is -2147483648
PASS output[11] is 2147483520
PASS output[12] is -65504
PASS output[13] is -2147483648
PASS output[14] is 65504
PASS output[15] is 2147483520
PASS output[16] is 0
PASS output[17] is 0
PASS output[18] is -65504
PASS output[19] is 65504
PASS output[20] is 0
PASS output[21] is 0
PASS output[22] is -2147483648
PASS output[23] is 2147483520
PASS output2[0] is 0
PASS output2[1] is 0
PASS output2[2] is 65504
PASS output2[3] is 4294967040
PASS output2[4] is 0
PASS output2[5] is 0
PASS output2[6] is 0
PASS output2[7] is 65504
PASS output2[8] is 0
PASS output2[9] is 0
PASS output2[10] is 0
PASS output2[11] is 4294967040
PASS output2[12] is 0
PASS output2[13] is 0
PASS output2[14] is 65504
PASS output2[15] is 4294967040
PASS output2[16] is 0
PASS output2[17] is 0
PASS output2[18] is 0
PASS output2[19] is 65504
PASS output2[20] is 0
PASS output2[21] is 0
PASS output2[22] is 0
PASS output2[23] is 4294967040
